refactor: combine analysis API endpoints to one file
